Response Entity

+++ src/main/java/com/pixel/movies/service/MovieService.java
@@ -8,6 +8,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import javax.validation.constraints.Null;
 import java.util.List;
 
 @Service
@@ -16,37 +17,58 @@ public class MovieService {
     @Autowired
     private MovieRepository movieRepository;
 
-    public Movie createMovie(Movie movie) {
-
-        return validateMovie(movie);
-    }
-
-    public String deleteMovie(Long id) {
-        movieRepository.deleteById(id);
-        return "Movie " + id + " was deleted";
-    }
-
-    public Movie getMovieById(Long id) {
-        return movieRepository.findById(id).get();
-    }
-
-    public List<Movie> getAll() {
-        return movieRepository.findAll();
-    }
-
-    public Movie updateMovieById(Movie movie, Long id) {
-        movie.setId(id);
-        return movieRepository.save(movie);
-    }
-
-    private Movie validateMovie(Movie movie) {
-        List<Movie> movies = getAll();
-        for (int i = 0; i < movies.size(); i++) {
-            if (movies.get(i).getTitle().equals(movie.getTitle())) {
-                throw new RuntimeException();
-            }
+    public ResponseEntity<String> createMovie(Movie movie) {
+        if (validateIfMovieExistsByTitle(movie)) {
+            return new ResponseEntity<>("This title already exists", HttpStatus.CONFLICT);
+        } else {
+            movieRepository.save(movie);
+            return new ResponseEntity<>("The movie was created", HttpStatus.CREATED);
         }
-        return movieRepository.save(movie);
+    }
+
+    public ResponseEntity<String> deleteMovie(Long id) {
+        if (movieRepository.existsById(id)) {
+            movieRepository.deleteById(id);
+            return new ResponseEntity<>("Movie " + id + " was deleted", HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>("Id " + id + " was not found", HttpStatus.NOT_FOUND);
+        }
+    }
+
+    public ResponseEntity<Movie> getMovieById(Long id) {
+        if (movieRepository.existsById(id)) {
+            return new ResponseEntity<>(movieRepository.findById(id).get(), HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+
+    }
+
+    public ResponseEntity<List<Movie>> getAll() {
+        if (movieRepository.count() <= 0) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        } else {
+            return new ResponseEntity<>(movieRepository.findAll(), HttpStatus.OK);
+        }
+
+    }
+
+    public ResponseEntity<String> updateMovieById(Movie movie, Long id) {
+        if (movieRepository.existsById(id)) {
+            if (validateIfMovieExistsByTitle(movie)) {
+                return new ResponseEntity<>("Movie already exists", HttpStatus.CONFLICT);
+            } else {
+                movie.setId(id);
+                movieRepository.save(movie);
+                return new ResponseEntity<>("Update was complete", HttpStatus.OK);
+            }
+        } else {
+            return new ResponseEntity<>("ID doesn't exist", HttpStatus.NOT_FOUND);
+        }
+    }
+
+    private Boolean validateIfMovieExistsByTitle(Movie movie) {
+        return movieRepository.existsByTitle(movie.getTitle());
     }
 }
